[REFACTOR/#170] image 패키지 분리
변경 내용
- common 패키지로 분리했던 image 관련 파일을 별도의 패키지로 분리
