+++ src/main/java/com/favoriteplace/app/item/domain/Item.java
@@ -1,6 +1,6 @@
 package com.favoriteplace.app.item.domain;
 
-import com.favoriteplace.app.common.domain.Image;
+import com.favoriteplace.app.image.domain.Image;
 import com.favoriteplace.app.common.domain.BaseTimeEntity;
 import com.favoriteplace.app.item.domain.enums.ItemCategory;
 import com.favoriteplace.app.item.domain.enums.ItemType;
