+++ src/main/java/com/favoriteplace/app/pilgrimage/domain/Pilgrimage.java
@@ -1,6 +1,6 @@
 package com.favoriteplace.app.pilgrimage.domain;
 
-import com.favoriteplace.app.common.domain.Image;
+import com.favoriteplace.app.image.domain.Image;
 import com.favoriteplace.app.common.domain.BaseTimeEntity;
 
 import com.favoriteplace.app.rally.domain.Address;
